gestion productos

+++ FormsApp/WinForms_Usuarios/Asignacion_Rol.cs
@@ -18,6 +18,7 @@ namespace FormsApp.WinForms_Usuarios
 
         Service_User_Rol.Gestion_User_RolClient objwcf = new Service_User_Rol.Gestion_User_RolClient();
         DataSet dataSet = new DataSet();
+        static int column = 0;
         public Asignacion_Rol()
         {
             InitializeComponent();
@@ -40,12 +41,21 @@ namespace FormsApp.WinForms_Usuarios
             var data = objwcf.List_User_rol();
             Type type = dataSet.GetType();
             var dataset = Deserialize(data, type);
-            roles.DataSource = dataset.Tables["list_roles"];
-            roles.DisplayMember = "nombre";
-            roles.ValueMember = "codigo"; 
-
+            if(dataset.Tables["list_roles"].Rows.Count != 0)
+            {
+                roles.DataSource = dataset.Tables["list_roles"];
+                roles.DisplayMember = "nombre";
+                roles.ValueMember = "codigo";
+            }
+            else
+            {
+                roles.DataSource = null;
+            }
+           
+            
             list_user_rol.DataSource = dataset.Tables["list_User"];
-            list_user_rol.Columns[0].Visible = false;
+            column = list_user_rol.Columns.Count;
+            //list_user_rol.Columns[0].Visible = false;
         }
         private void Asignacion_Rol_Load(object sender, EventArgs e)
         {
@@ -56,10 +66,27 @@ namespace FormsApp.WinForms_Usuarios
 
         private void add_Rol_Click(object sender, EventArgs e)
         {
-            var id_rol = Convert.ToInt32(roles.SelectedValue.ToString());
-            var id_user = Convert.ToInt32(list_user_rol.CurrentRow.Cells[0].Value.ToString());
-            objwcf.insert_rol_user(id_rol, id_user);
-            loadlist_Roles_user();
+
+            if (column > 1)
+            {
+
+                if (roles.Items.Count != 0)
+                {
+
+                    var id_rol = Convert.ToInt32(roles.SelectedValue.ToString());
+                    var id_user = Convert.ToInt32(list_user_rol.CurrentRow.Cells[0].Value.ToString());
+                    objwcf.insert_rol_user(id_rol, id_user);
+                    loadlist_Roles_user();
+                }
+                
+                
+                //var id_rol = Convert.ToInt32(roles.SelectedValue.ToString());
+                //var id_user = Convert.ToInt32(list_user_rol.CurrentRow.Cells[0].Value.ToString());
+                //objwcf.insert_rol_user(id_rol, id_user);
+                //loadlist_Roles_user();
+
+            }
+            //MessageBox.Show(list_user_rol.CurrentRow.Cells[0].Value.ToString());
 
 
         }
@@ -70,10 +97,19 @@ namespace FormsApp.WinForms_Usuarios
         {
 
 
-            var id_rol = Convert.ToInt32(roles.SelectedValue.ToString());
-            var id_user = Convert.ToInt32(list_user_rol.CurrentRow.Cells[0].Value);
-            objwcf.delete_rol_user(id_rol, id_user);
-            loadlist_Roles_user();
+            if (column > 1)
+            {
+
+                if (roles.Items.Count != 0)
+                {
+                    var id_rol = Convert.ToInt32(roles.SelectedValue.ToString());
+                    var id_user = Convert.ToInt32(list_user_rol.CurrentRow.Cells[0].Value);
+                    objwcf.delete_rol_user(id_rol, id_user);
+                    loadlist_Roles_user();
+                }
+               
+            }
+           
         }
 
         private void button1_Click(object sender, EventArgs e)
@@ -96,10 +132,18 @@ namespace FormsApp.WinForms_Usuarios
 
         private void list_user_rol_CellClick(object sender, DataGridViewCellEventArgs e)
         {
+            //list_temas_rece.CurrentRow.Cells[4].Value != DBNull.Value
+            if(list_user_rol.Columns.Count > 3)
+            {
+                var _roles = list_user_rol.CurrentRow.Cells[2].Value.ToString();
+                roles.Text = _roles;
+            }
 
-            var data = list_user_rol.CurrentRow.Cells[2].Value.ToString();
-            roles.Text = data;
+            //if (list_user_rol.CurrentRow.Cells[2].Value != DBNull.Value)
+            //{
 
+                
+            //}
             //MessageBox.Show(roles.SelectedValue.ToString());
 
             //int codigo = (int)((DataRowView)roles.Items[roles.SelectedIndex])["codigo"];
